requirements.txt
uvicorn==0.34.2
fastapi==0.115.12
requests==2.32.4
python-dotenv==1.1.0